django/requirements.txt
boto3==1.2.3
ipython==3.0.0
django-extensions==1.5.9
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
django-filter==0.13.0
factory_boy==2.7.0
